Organize repo : add 'pages'
